Stop Client.connected() emitting unnecessarily
I found that I was seeing values emitted on the Client.connected()
observable when a peer was connecting but not yet connected.

A connecting peer does not change the `connectedPeers` property of
the `MCSession`, but it does cause a call to the
`session(_:peer:didChange:)` method on the delegate. This means that the
delegate proxy will set the value of `_connected` even when there is no
change to the `connectedPeers` value it passes.

This change simply checks if the `state` of the peer is `.connecting`.
If it is not (ie. it is `connected` or `disconnected`) then we update
the value of `_connected` (and thus `connected()`).